Move TimeAgo to main component folder
